重新设计文章实体对象

+++ pojo/src/main/java/top/tobycold/dao/SummaryPages.java
@@ -10,5 +10,5 @@ import java.util.List;
 @Builder
 public class SummaryPages {
     List<SummaryEntity> summaryEntities;
-    Long total;
+    int total;
 }
